requirements.txt
python-telegram-bot==20.7
requests
vk_api
python-dotenv
beautifulsoup4